LayoutTests/imported/w3c/web-platform-tests/FileAPI/url/multi-global-origin-serialization.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/FileAPI/support/incumbent.sub.html
CONSOLE MESSAGE: TypeError: frames[0].createBlobURL is not a function. (In 'frames[0].createBlobURL()', 'frames[0].createBlobURL' is undefined)


Harness Error (FAIL), message = TypeError: frames[0].createBlobURL is not a function. (In 'frames[0].createBlobURL()', 'frames[0].createBlobURL' is undefined)
